pi_scanner/requirements.txt
pyjwt==2.0.1
paho-mqtt==1.5.1
bleak
cryptography